+++ src/Testcontainers/Configurations/Modules/Databases/MySqlTestcontainerConfiguration.cs
@@ -1,5 +1,6 @@
 namespace DotNet.Testcontainers.Configurations
 {
+  using System;
   using DotNet.Testcontainers.Builders;
   using JetBrains.Annotations;
 
@@ -7,11 +8,11 @@ namespace DotNet.Testcontainers.Configurations
   [PublicAPI]
   public class MySqlTestcontainerConfiguration : TestcontainerDatabaseConfiguration
   {
+    private const string MySqlRootUsername = "root";
+
     private const string MySqlImage = "mysql:8.0.28";
 
     private const int MySqlPort = 3306;
-
-    private const string RootUsername = "root";
 
     /// <summary>
     /// Initializes a new instance of the <see cref="MySqlTestcontainerConfiguration" /> class.
@@ -43,23 +44,17 @@ namespace DotNet.Testcontainers.Configurations
     {
       get
       {
-        string username;
-        if (this.Environments.TryGetValue("MYSQL_USER", out username))
-        {
-          return username;
-        }
-        else
-        {
-          return RootUsername;
-        }
+        return this.Environments.TryGetValue("MYSQL_USER", out var username) ? username : MySqlRootUsername;
       }
+
       set
       {
-        // Only set the username if it is not "root", as mysql does not allow it.
-        if (value != RootUsername)
+        if (MySqlRootUsername.Equals(value, StringComparison.OrdinalIgnoreCase))
         {
-          this.Environments["MYSQL_USER"] = value;
+          return;
         }
+
+        this.Environments["MYSQL_USER"] = value;
       }
     }
 
@@ -69,8 +64,8 @@ namespace DotNet.Testcontainers.Configurations
       get => this.Environments["MYSQL_PASSWORD"];
       set
       {
-        this.Environments["MYSQL_ROOT_PASSWORD"] = value;
         this.Environments["MYSQL_PASSWORD"] = value;
+        this.Environments["MYSQL_ROOT_PASSWORD"] = value;
       }
     }