+++ tests/Testcontainers.Tests/Fixtures/Containers/Unix/Modules/Databases/MySqlRootUsernameFixture.cs
@@ -1,0 +1,14 @@
+﻿namespace DotNet.Testcontainers.Tests.Fixtures
+{
+  using DotNet.Testcontainers.Configurations;
+  using JetBrains.Annotations;
+
+  [UsedImplicitly]
+  public sealed class MySqlRootUsernameFixture : MySqlFixture
+  {
+    public MySqlRootUsernameFixture()
+      : base(new MySqlTestcontainerConfiguration { Database = "db", Username = "root", Password = "root" })
+    {
+    }
+  }
+}